LRAC-12472 Use 1 shard count by default

+++ modules/dxp/apps/osb/osb-asah/osb-asah-dataflow-java/src/main/java/com/liferay/osb/asah/dataflow/ingestion/dxp/DXPEntitiesIngestionPipelineOptions.java
@@ -35,7 +35,7 @@ public interface DXPEntitiesIngestionPipelineOptions
 	@Validation.Required
 	public String getPubsubSubscription();
 
-	@Default.Integer(4)
+	@Default.Integer(1)
 	@Description("Return the shard count")
 	public int getShardCount();
